refactor(analytics-highlights) get-highlights when render tab page
