+++ src/components/KeyResult/CheckInForm/form.tsx
@@ -10,7 +10,7 @@ import activeAndOwnedByUserQuery from 'src/components/KeyResult/ActiveAndOwnedBy
 import { KeyResult, KeyResultCheckIn } from 'src/components/KeyResult/types'
 import notificationsQuery from 'src/components/Notifications/Modal/queries.gql'
 import accordionItemPanelQuery from 'src/components/Objective/Accordion/Item/Panel/queries.gql'
-import GET_KEY_RESULTS_HIGHLIGHTS from 'src/components/Page/Team/Highlights/hooks/getKeyRusultsHighlights/get-key-results-highlights.gql'
+import GET_KEY_RESULTS_HIGHLIGHTS from 'src/components/Page/Team/Highlights/get-key-results-highlights.gql'
 import { EventType } from 'src/state/hooks/useEvent/event-type'
 import { useEvent } from 'src/state/hooks/useEvent/hook'
 import { keyResultCheckInCommentEnabled } from 'src/state/recoil/key-result/check-in'
