+++ src/components/KeyResult/Single/Sections/Owner/user-list.tsx
@@ -2,7 +2,7 @@ import { useMutation } from '@apollo/client'
 import React from 'react'
 import { useRecoilState, useRecoilValue } from 'recoil'
 
-import GET_KEY_RESULTS_HIGHLIGHTS from 'src/components/Page/Team/Highlights/hooks/getKeyRusultsHighlights/get-key-results-highlights.gql'
+import GET_KEY_RESULTS_HIGHLIGHTS from 'src/components/Page/Team/Highlights/get-key-results-highlights.gql'
 import GET_NO_RELATED_MEMBERS from 'src/components/Page/Team/Highlights/hooks/getNoRelatedMembers/get-no-related-members.gql'
 import { selectedTeamIdHighlight } from 'src/state/recoil/team/highlight/selected-team-id-highlight'
