+++ src/components/Page/Team/Highlights/get-key-results-highlights.gql
