--- src/components/Page/Team/Highlights/hooks/getKeyRusultsHighlights/index.ts
@@ -1,1 +1,0 @@
-export { useGetKeyResultsHighlights } from './get-key-results-highlights'
